requirements.txt
flet
SpeechRecognition
pyttsx3
pyaudio
setuptools
pocketsphinx
python-dotenv
flet_route